Fixing the string outputs for the new python version requirement.
This also adds a note to the Readme about the change. It isn't
actually a hard requirement and other versions should still work just
fine, including with the old version of the strings. The problem only
arises when trying to run the tests as they are checking for exact
equality. If developing locally on a different python version and the
tests fail, they are okay to ignore as long as there aren't actually
any changes that would change the output for these strings. If there
are changes, copy the output in the pipeline and copy it to the tests.